move types from eventCard to event folder root
