+++ src/domain/event/eventCard/EventCard.tsx
@@ -18,9 +18,13 @@ import {
   getEventPrice,
   isEventClosed,
 } from '../EventUtils';
-import { EventFields } from '../types';
+import {
+  EVENT_ROUTE_MAPPER,
+  EventFields,
+  EVENTS_ROUTE_MAPPER,
+  EventType,
+} from '../types';
 import styles from './eventCard.module.scss';
-import { EVENT_ROUTE_MAPPER, EVENTS_ROUTE_MAPPER, EventType } from './types';
 import { addPlaceFromPathToQueryString } from './utils';
 
 interface Props {
