+++ src/domain/event/types.ts
@@ -2,12 +2,26 @@ import {
   CourseFieldsFragment,
   EventFieldsFragment,
 } from '../../generated/graphql';
+import { ROUTES } from '../app/routes/constants';
 
 export type KeywordOption = {
   id: string;
   name: string;
 };
 
-export type EventType = 'course' | 'event';
-
 export type EventFields = EventFieldsFragment | CourseFieldsFragment;
+
+export enum EventType {
+  EVENT = 'event',
+  COURSE = 'course',
+}
+
+export const EVENT_ROUTE_MAPPER = {
+  [EventType.EVENT]: ROUTES.EVENT,
+  [EventType.COURSE]: ROUTES.COURSE,
+};
+
+export const EVENTS_ROUTE_MAPPER = {
+  [EventType.EVENT]: ROUTES.EVENTS,
+  [EventType.COURSE]: ROUTES.COURSES,
+};
