+++ src/domain/eventSearch/EventSearchPageContainer.tsx
@@ -14,8 +14,8 @@ import useIsSmallScreen from '../../hooks/useIsSmallScreen';
 import useLocale from '../../hooks/useLocale';
 import MainContent from '../app/layout/MainContent';
 import PageWrapper from '../app/layout/PageWrapper';
-import { EventType } from '../event/eventCard/types';
 import { getLargeEventCardId } from '../event/EventUtils';
+import { EventType } from '../event/types';
 import EventList from '../eventList/EventList';
 import LandingPageMeta from '../landingPage/landingPageMeta/LandingPageMeta';
 import { isLanguageSupported } from '../landingPage/utils';
